Fix verify_scancnf_results
In verify_scancnf_results, the status of the MLME-SCAN.confirm must be
modified if all entries were removed from the results list. In the case
of a scanning device with macAutoRequest=false, the results list will be
empty even if beacons were received during the scan. This function would
have erroneously set the status to "NO_BEACON" when it should remain
"SUCCESS".